backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
google-cloud-storage==3.3.0
google-auth==2.40.3
google-cloud-speech==2.33.0
soundfile==0.13.1
librosa==0.11.0
cohere==5.17.0
python-pptx==0.6.23
spire.presentation==8.12.0
PyPDF2==3.0.1
python-multipart==0.0.9
Pillow==10.0.0
